refactor: checklists를 private, shared폴더로 분리.
