+++ server/src/folders/entities/folder.entity.ts
@@ -1,6 +1,6 @@
 import { BaseModel } from '../../common/entity/base.entity';
 import { Column, Entity, ManyToOne, OneToMany } from 'typeorm';
-import { PrivateChecklistModel } from '../../checklists/entities/private-checklist.entity';
+import { PrivateChecklistModel } from '../../private-checklists/entities/private-checklist.entity';
 import { UserModel } from '../../users/entities/user.entity';
 
 @Entity()
