+++ server/src/shared-checklists/entities/shared-checklist.entity.ts
