+++ server/src/users/entities/user.entity.ts
@@ -1,7 +1,7 @@
 import { BaseModel } from 'src/common/entity/base.entity';
 import { Column, Entity, Generated, ManyToMany, OneToMany } from 'typeorm';
-import { PrivateChecklistModel } from '../../checklists/entities/private-checklist.entity';
-import { SharedChecklistModel } from '../../checklists/entities/shared-checklist.entity';
+import { PrivateChecklistModel } from '../../private-checklists/entities/private-checklist.entity';
+import { SharedChecklistModel } from '../../shared-checklists/entities/shared-checklist.entity';
 import { FolderModel } from '../../folders/entities/folder.entity';
 
 @Entity()
